PKG -- [sdk] Move createSignableVoucher to separate module
