+++ packages/sdk/src/resolve/resolve-signatures.js
@@ -1,9 +1,14 @@
 import {isTransaction} from "../interaction/interaction.js"
-import {sansPrefix, withPrefix} from "@onflow/util-address"
+import {sansPrefix} from "@onflow/util-address"
 import {
   encodeTransactionPayload as encodeInsideMessage,
   encodeTransactionEnvelope as encodeOutsideMessage,
 } from "../encode/encode.js"
+import {
+  createSignableVoucher,
+  findInsideSigners,
+  findOutsideSigners,
+} from "./voucher.js"
 
 export async function resolveSignatures(ix) {
   if (isTransaction(ix)) {
@@ -30,20 +35,6 @@ export async function resolveSignatures(ix) {
   return ix
 }
 
-export function findInsideSigners(ix) {
-  // Inside Signers Are: (authorizers + proposer) - payer
-  let inside = new Set(ix.authorizations)
-  inside.add(ix.proposer)
-  inside.delete(ix.payer)
-  return Array.from(inside)
-}
-
-function findOutsideSigners(ix) {
-  // Outside Signers Are: (payer)
-  let outside = new Set([ix.payer])
-  return Array.from(outside)
-}
-
 function fetchSignature(ix, payload) {
   return async function innerFetchSignature(id) {
     const acct = ix.accounts[id]
@@ -55,47 +46,6 @@ function fetchSignature(ix, payload) {
     //   ix.accounts[id].keyId = keyId
     // }
     ix.accounts[id].signature = signature
-  }
-}
-
-export const createSignableVoucher = ix => {
-  const buildAuthorizers = () => {
-    const authorizations = ix.authorizations
-      .map(cid => withPrefix(ix.accounts[cid].addr))
-      .reduce((prev, current) => {
-        return prev.find(item => item === current) ? prev : [...prev, current]
-      }, [])
-    return authorizations[0] ? authorizations : []
-  }
-
-  const buildInsideSigners = () =>
-    findInsideSigners(ix).map(id => ({
-      address: withPrefix(ix.accounts[id].addr),
-      keyId: ix.accounts[id].keyId,
-      sig: ix.accounts[id].signature,
-    }))
-
-  const buildOutsideSigners = () =>
-    findOutsideSigners(ix).map(id => ({
-      address: withPrefix(ix.accounts[id].addr),
-      keyId: ix.accounts[id].keyId,
-      sig: ix.accounts[id].signature,
-    }))
-
-  return {
-    cadence: ix.message.cadence,
-    refBlock: ix.message.refBlock || null,
-    computeLimit: ix.message.computeLimit,
-    arguments: ix.message.arguments.map(id => ix.arguments[id].asArgument),
-    proposalKey: {
-      address: withPrefix(ix.accounts[ix.proposer].addr),
-      keyId: ix.accounts[ix.proposer].keyId,
-      sequenceNum: ix.accounts[ix.proposer].sequenceNum,
-    },
-    payer: withPrefix(ix.accounts[ix.payer].addr),
-    authorizers: buildAuthorizers(),
-    payloadSigs: buildInsideSigners(),
-    envelopeSigs: buildOutsideSigners(),
   }
 }
