+++ packages/sdk/src/sdk.js
@@ -31,7 +31,7 @@ export {
   isGetBlockHeader,
   isGetCollection,
 } from "./interaction/interaction.js"
-export {createSignableVoucher} from "./resolve/resolve-signatures"
+export {createSignableVoucher} from "./resolve/voucher.js"
 export {encodeMessageFromSignable} from "./wallet-utils/encode-signable.js"
 export {template as cadence} from "@onflow/util-template"
 export {template as cdc} from "@onflow/util-template"
